services/web/requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
gunicorn==20.1.0
mysql-connector-python==8.0.32